+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 ## Change log
 
-### 7.8.1 (08/19/2021 - )
+### 7.8.1 (08/19/2021 - 09/14/2021)
 
 * db: batchInsert returns Optional<long[]> for auto incremental PK
 * db: update mysql driver to 8.0.26
 * httpClient: support client ssl auth
 * site: removed Session.timeout(Duration), it proved not useful, for app level remember me, it's better handle in app level with custom token handling
 * redis: support password auth for redis/cache
+* ws: added WebContext.responseCookie to allow WS to assign cookie to response
 
 ### 7.8.0 (08/04/2021 - 08/17/2021)   !!! breaking changes, pls read details